+++ lib/changeset.ex
@@ -49,16 +49,20 @@ defmodule Formex.Ecto.Changeset do
     Form.get_fields(form)
     |> Enum.filter(&(&1.type == :multiple_select))
     |> Enum.reduce(changeset, fn field, changeset ->
-      module = form.struct_module.__schema__(:association, field.name).related
-      ids    = form.mapped_params[to_string(field.name)] || []
+      case form.struct_module.__schema__(:association, field.name) do
+        nil ->
+          changeset
+        module ->
+          ids = form.mapped_params[to_string(field.name)] || []
 
-      associated = module
-      |> where([c], c.id in ^ids)
-      |> @repo.all
-      |> Enum.map(&Ecto.Changeset.change/1)
+          associated = module.related
+          |> where([c], c.id in ^ids)
+          |> @repo.all
+          |> Enum.map(&Ecto.Changeset.change/1)
 
-      changeset
-      |> put_assoc(field.name, associated)
+          changeset
+          |> put_assoc(field.name, associated)
+      end
     end)
   end